+++ src/__fixtures__/all-of.json
