+++ src/__fixtures__/nested-properties.json
