+++ src/__fixtures__/properties.json
