+++ src/components/Types.tsx
@@ -1,7 +1,7 @@
 import { JSONSchema4TypeName } from 'json-schema';
 import * as React from 'react';
 import { MutedText } from '../common/MutedText';
-import { JSONSchema4CombinerName } from '../renderers/types';
+import { JSONSchema4CombinerName } from '../types';
 import { Type } from './Type';
 
 interface ITypes {
