+++ src/types.ts
@@ -1,55 +1,51 @@
-import { Dictionary, ISchema, JsonPath } from '@stoplight/types';
-import { ICustomTheme } from '@stoplight/ui-kit';
-import { ReactNodeArray } from 'react';
+import { Dictionary, JsonPath } from '@stoplight/types';
+import { JSONSchema4, JSONSchema4TypeName } from 'json-schema';
 
-export interface ICommonProps {
-  originalSchema?: ISchema;
-  schema?: ISchema;
-  defaultExpandedDepth: number;
-  prop?: ISchema;
-  parentName?: string;
-  rowElems: ReactNodeArray;
-  expandedRows: Dictionary<boolean>;
+export const enum SchemaKind {
+  Any = 'any',
+  String = 'string',
+  Number = 'number',
+  Integer = 'integer',
+  Boolean = 'boolean',
+  Null = 'null',
+  Array = 'array',
+  Object = 'object',
+}
+
+export type JSONSchema4CombinerName = 'allOf' | 'anyOf' | 'oneOf';
+
+export type JSONSchema4Annotations = 'title' | 'description' | 'default' | 'examples';
+
+export interface ICombinerNode {
+  readonly combiner: JSONSchema4CombinerName;
+  properties?: JSONSchema4[];
+  annotations: Pick<JSONSchema4, JSONSchema4Annotations>;
+}
+
+export interface IBaseNode {
+  readonly type: JSONSchema4TypeName | JSONSchema4TypeName[];
+  annotations: Pick<JSONSchema4, JSONSchema4Annotations>;
+  validations: Dictionary<unknown>;
+}
+
+export interface IArrayNode extends IBaseNode, Pick<JSONSchema4, 'items'> {}
+
+export interface IObjectNode extends IBaseNode, Pick<JSONSchema4, 'properties' | 'patternProperties'> {}
+
+export interface IObjectPropertyNode extends IBaseNode {
+  name: string;
+}
+
+export type SchemaNode = ICombinerNode | IBaseNode | IArrayNode | IObjectNode | IObjectPropertyNode;
+
+export type SchemaTreeNode = SchemaNode & {
+  level: number;
   path: JsonPath;
-  propName?: string;
-  required?: boolean | string[];
-  hideInheritedFrom?: boolean;
-  hideRoot?: boolean;
+  showDivider?: boolean;
+  subtype?: IBaseNode['type'];
+  expanded?: boolean;
+  required?: boolean;
+  inheritedFrom?: string;
+};
 
-  toggleExpandRow(rowKey: string, expanded: boolean): void;
-}
 
-export interface IJSONSchemaViewerTheme extends ICustomTheme {
-  canvas: {
-    bg: string;
-    fg: string;
-    error: string;
-    muted: string;
-  };
-
-  row: {
-    hoverFg?: string;
-    hoverBg: string;
-    evenFg?: string;
-    evenBg: string;
-  };
-
-  types: {
-    array: string;
-    object: string;
-    allOf: string;
-    oneOf: string;
-    anyOf: string;
-    null: string;
-    integer: string;
-    number: string;
-    boolean: string;
-    binary: string;
-    string: string;
-    $ref: string;
-  };
-
-  divider: {
-    bg: string;
-  };
-}
