+++ src/utils/isCombiner.ts
@@ -1,3 +1,3 @@
-import { ICombinerNode, SchemaNode } from '../renderers/types';
+import { ICombinerNode, SchemaNode } from '../types';
 
 export const isCombiner = (node: SchemaNode): node is ICombinerNode => 'combiner' in node;
